更改

+++ guava/src/com/google/common/collect/Hashing.java
@@ -31,12 +31,8 @@ import javax.annotation.Nullable;
 final class Hashing {
   private Hashing() {}
 
-  /*
-   * These should be ints, but we need to use longs to force GWT to do the multiplications with
-   * enough precision.
-   */
-  private static final long C1 = 0xcc9e2d51;
-  private static final long C2 = 0x1b873593;
+  private static final int C1 = 0xcc9e2d51;
+  private static final int C2 = 0x1b873593;
 
   /*
    * This method was rewritten in Java from an intermediate step of the Murmur hash function in
@@ -47,7 +43,7 @@ final class Hashing {
    * hereby disclaims copyright to this source code.
    */
   static int smear(int hashCode) {
-    return (int) (C2 * Integer.rotateLeft((int) (hashCode * C1), 15));
+    return C2 * Integer.rotateLeft(hashCode * C1, 15);
   }
 
   static int smearedHash(@Nullable Object o) {
